Generative AI/MCPLabs/mcp_chat_app/requirements.txt
asyncio
paho-mqtt
